make logging its own module
